requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
httpx==0.27.0
gspread==6.1.2
google-auth==2.29.0
python-dotenv==1.0.1
requests==2.31.0
